refactor(server): consolidate module directories
