+++ src/server/service/query.h
@@ -9,7 +9,7 @@
 #include "semantic/relation_kind.h"
 #include "semantic/symbol_kind.h"
 #include "server/protocol/agentic.h"
-#include "server/workspace/workspace.h"
+#include "server/state/workspace.h"
 
 #include "kota/ipc/lsp/position.h"
 #include "kota/ipc/lsp/protocol.h"
