+++ src/server/state/config.cpp
@@ -1,4 +1,4 @@
-#include "server/workspace/config.h"
+#include "server/state/config.h"
 
 #include <algorithm>
