+++ src/server/state/file_tracker.cpp
@@ -1,4 +1,4 @@
-#include "server/workspace/file_tracker.h"
+#include "server/state/file_tracker.h"
 
 #include <algorithm>
 #include <chrono>
