+++ src/server/state/session_store.cpp
@@ -1,4 +1,4 @@
-#include "server/session/session_store.h"
+#include "server/state/session_store.h"
 
 #include <type_traits>
 #include <utility>
