+++ src/server/transport/agentic.cpp
@@ -1,4 +1,4 @@
-#include "server/service/agentic.h"
+#include "server/transport/agentic.h"
 
 #include <memory>
 #include <print>
