+++ src/server/transport/master_server.cpp
@@ -1,4 +1,4 @@
-#include "server/service/master_server.h"
+#include "server/transport/master_server.h"
 
 #include <list>
 #include <memory>
@@ -6,9 +6,9 @@
 #include <vector>
 
 #include "server/protocol/worker.h"
-#include "server/service/agent_client.h"
-#include "server/service/lsp_client.h"
-#include "server/workspace/file_tracker.h"
+#include "server/transport/agent_client.h"
+#include "server/transport/lsp_client.h"
+#include "server/state/file_tracker.h"
 #include "support/anomaly.h"
 #include "support/filesystem.h"
 #include "support/logging.h"
