+++ src/server/transport/master_server.h
@@ -6,16 +6,16 @@
 #include <vector>
 
 #include "server/compiler/compiler.h"
-#include "server/context/context_resolver.h"
-#include "server/feature/feature_router.h"
-#include "server/index/background_indexer.h"
-#include "server/index/query.h"
-#include "server/session/session.h"
-#include "server/session/session_store.h"
+#include "server/compiler/context_resolver.h"
+#include "server/service/feature_router.h"
+#include "server/compiler/background_indexer.h"
+#include "server/service/query.h"
+#include "server/state/session.h"
+#include "server/state/session_store.h"
 #include "server/worker/worker_pool.h"
-#include "server/workspace/config.h"
-#include "server/workspace/invalidator.h"
-#include "server/workspace/workspace.h"
+#include "server/state/config.h"
+#include "server/state/invalidator.h"
+#include "server/state/workspace.h"
 
 #include "kota/async/async.h"
 #include "kota/deco/deco.h"
