+++ tests/unit/server/file_tracker_tests.cpp
@@ -1,7 +1,7 @@
 #include "test/cdb_helper.h"
 #include "test/temp_dir.h"
 #include "test/test.h"
-#include "server/workspace/file_tracker.h"
+#include "server/state/file_tracker.h"
 
 namespace clice::testing {
 namespace {
