+++ tests/unit/server/rank_hosts_tests.cpp
@@ -1,5 +1,5 @@
 #include "test/test.h"
-#include "server/workspace/workspace.h"
+#include "server/state/workspace.h"
 
 namespace clice::testing {
 namespace {
